fork-dijstra: Instalando maven e as dependencias necessarias.
