+++ src/main/java/dijkstra/DefaultGraph.java
@@ -3,32 +3,32 @@ package dijkstra;
 import org.jgrapht.graph.DefaultWeightedEdge;
 import org.jgrapht.graph.WeightedPseudograph;
 
+import java.util.Arrays;
+import java.util.List;
+
 public class DefaultGraph {
 	private WeightedPseudograph<String, DefaultWeightedEdge> graph;
 	public DefaultGraph() {
 		graph = new WeightedPseudograph<>(DefaultWeightedEdge.class);
-		
+
 		generateStandardGraph();
     }
 
-	
 	public WeightedPseudograph<String, DefaultWeightedEdge> getDefaultGraph() {
 		return graph;
 	}
-	
+
 	private void generateStandardGraph() {
-		for (String vertice: getVerticeValues()) {
-			graph.addVertex(vertice);
-		}
-		
+		getVerticeValues().forEach(graph::addVertex);
+
 		String[][] edges = getEdgeValues();
-		
+
 		for (int i = 0; i < edges.length; i++) {
 			DefaultWeightedEdge edge = graph.addEdge(edges[i][0], edges[i][1]);
 			graph.setEdgeWeight(edge, Integer.parseInt(edges[i][2]));
 		}
 	}
-	
+
 	private String[][] getEdgeValues() {
 		return new String[][] {
 				{"A", "C", "1"},
@@ -56,8 +56,8 @@ public class DefaultGraph {
 				{"M", "P", "8"},
 		};
 	}
-	
-	private String[] getVerticeValues() {
-		return new String[] {"A", "B", "C", "D", "E", "F", "G", "H", "I", "J", "K", "L", "M", "N", "O", "P"};
+
+	private List<String> getVerticeValues() {
+		return Arrays.asList("A", "B", "C", "D", "E", "F", "G", "H", "I", "J", "K", "L", "M", "N", "O", "P");
 	}
 }
