--- src/module-info.java
@@ -1,10 +1,0 @@
-/**
- * 
- */
-/**
- * @author iasmim
- *
- */
-module dijkstra {
-	requires org.jgrapht.core;
-}
